refactor(nc-gui): remove nested page in user path
